# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampusEats API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all active vendors
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorBody"
      responses:
        "201":
          description: Created vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    put:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVendorBody"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}/stats:
    get:
      operationId: getVendorStats
      tags: [vendors]
      summary: Get vendor earnings and stats
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorStats"
  /vendors/{id}/orders:
    get:
      operationId: getVendorOrders
      tags: [vendors]
      summary: Get all orders for a vendor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, paid, preparing, ready, completed, cancelled]
      responses:
        "200":
          description: Vendor orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - in: query
          name: student_id
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
  /orders/student/{studentId}/recent:
    get:
      operationId: getStudentRecentOrders
      tags: [orders]
      summary: Get recent orders for a student (for "Order again" feature)
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (register)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: User session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, vendor, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role]
    Vendor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        stallName:
          type: string
        description:
          type: string
        mpesaNumber:
          type: string
        isActive:
          type: boolean
        imageUrl:
          type: string
        location:
          type: string
        pickupTimeMin:
          type: integer
        pickupTimeMax:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, stallName, isActive]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        vendorId:
          type: integer
        studentName:
          type: string
        status:
          type: string
          enum: [pending, paid, preparing, ready, completed, cancelled]
        totalAmount:
          type: number
        pickupCode:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, studentId, vendorId, status, totalAmount]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [id, orderId, menuItemId, menuItemName, quantity, unitPrice]
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            vendor:
              $ref: "#/components/schemas/Vendor"
...
    VendorStats:
      type: object
      properties:
        totalOrders:
          type: integer
        ordersToday:
          type: integer
        totalEarnings:
          type: number
        earningsToday:
          type: number
        pendingOrders:
          type: integer
        completedOrders:
          type: integer
      required: [totalOrders, ordersToday, totalEarnings, earningsToday, pendingOrders, completedOrders]
...
    CreateVendorBody:
      type: object
      properties:
        userId:
          type: integer
        stallName:
          type: string
        description:
          type: string
        mpesaNumber:
          type: string
        imageUrl:
          type: string
        location:
          type: string
        pickupTimeMin:
          type: integer
        pickupTimeMax:
          type: integer
      required: [stallName, mpesaNumber]
    UpdateVendorBody:
      type: object
      properties:
        stallName:
          type: string
        description:
          type: string
        mpesaNumber:
          type: string
        imageUrl:
          type: string
        location:
          type: string
        isActive:
          type: boolean
        pickupTimeMin:
          type: integer
        pickupTimeMax:
          type: integer
    CreateOrderBody:
      type: object
      properties:
        studentId:
          type: integer
        vendorId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
            required: [menuItemId, quantity]
        notes:
          type: string
        studentName:
          type: string
      required: [vendorId, items]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, preparing, ready, completed, cancelled]
      required: [status]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, vendor, admin]
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]