# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CampusEats API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyItemsItem } from "./createOrderBodyItemsItem";

export interface CreateOrderBody {
  studentId?: number;
  vendorId: number;
  items: CreateOrderBodyItemsItem[];
  notes?: string;
  studentName?: string;
}